+++ examples/rails_app/app/processors/webhook_processor.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'cdc/core'
+require 'cdc/concurrent'
+
+# Example I/O-oriented CDC processor executed through cdc-concurrent.
+class WebhookProcessor < CDC::Core::Processor
+  concurrent_safe!
+
+  def process(event)
+    Rails.logger.info(
+      "downstream cdc event #{event.operation} #{event.qualified_table_name} primary_key=#{event.primary_key.inspect}"
+    )
+    CDC::Core::ProcessorResult.success(event)
+  end
+end